requirements.txt
disnake==2.9.2
pychrdb==1.2.0
aiohttp==3.8.5
emoji==2.7.0
Pillow==9.5.0
python-dotenv==1.0.0